pyproject.toml
[project]
name = "epd-ble-sender"
version = "0.1.0"
description = "A command-line tool to send images and text to an EPD via BLE."
authors = [
    {name = "Roo"},
]
dependencies = [
    "bleak>=0.21.1",
    "Pillow>=10.0.0",
    "click>=8.1.3",
    "numpy",
]
requires-python = ">=3.8"

[project.scripts]
epd-send = "epd_ble_sender.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["epd_ble_sender/"]